# lib/api-zod/src/generated/types/bggGeelist.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BggItem } from "./bggItem";

export interface BggGeelist {
  listTitle: string;
  totalItems: number;
  items: BggItem[];
}