# lib/api-zod/src/generated/types/bggItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BggItemStatus } from "./bggItemStatus";
import type { BggItemType } from "./bggItemType";

export interface BggItem {
  id: string;
  gameTitle: string;
  price: number;
  type: BggItemType;
  status: BggItemStatus;
  buyerSeller?: string;
  condition?: string;
  notes?: string;
}